Exit process when app stage gets closed
Previously this did only happen when the root stage was getting closed, which does not have an
actual window. On macOS this resembles the behavior of native apps but I wanted to quit when closing
the last application window which would be the expected behavior on Linux and Windows usually.